requirements.txt
requests
streamlit
python-dotenv
numpy
pandas